requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
drf-yasg
gunicorn
dj-database-url
whitenoise